Update lit and labs/ssr and split out libs
