+++ packages/labs/ssr-react/src/node/enable-lit-ssr.ts
@@ -10,19 +10,8 @@
  * Lit components.
  */
 
-// TODO(augustjk) Remove when https://github.com/lit/lit/pull/3522 lands
-import '@lit-labs/ssr/lib/install-global-dom-shim.js';
 import React from 'react';
-import {patchCreateElement} from '../lib/utils.js';
-
-// The base class of `ReactiveElement` could be an empty class from the Node
-// build of reactive-element if it was loaded before the DOM shim. Replace
-// prototype if that's the case.
-// TODO(augustjk) Remove when https://github.com/lit/lit/pull/3522 lands
-import {ReactiveElement} from 'lit';
-if (!('getAttribute' in ReactiveElement.prototype)) {
-  Object.setPrototypeOf(ReactiveElement.prototype, HTMLElement.prototype);
-}
+import {patchCreateElement} from '../lib/patch-create-element.js';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 (React.createElement as any) = patchCreateElement(React.createElement);
